extract class to encrypt/decrypt db on demand.
